For a given person, retrieve all relations where the person is involved, and all the relations between the other persons identified. On double click on a person, do the above and refresh the graph. Refactoring the code to handle the flags in Domain Value in a separate class. For existing Persons, display existing attribute values and allow user to modify. -- Not yet complete

+++ src/main/java/org/sakuram/relation/repository/AttributeValueRepository.java
@@ -3,6 +3,6 @@ package org.sakuram.relation.repository;
 import org.sakuram.relation.bean.AttributeValue;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface AttributeValueRepository extends JpaRepository<AttributeValue, Integer>{
+public interface AttributeValueRepository extends JpaRepository<AttributeValue, Long>{
 
 }
